requirements.txt
pandas
pydantic
sqlalchemy
httpx
aiolimiter
python-dotenv
pytest
boto3
pytest-asyncio
respx
python-dotenv
prefect==3.6.24
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
